Fix multiline text rendering not wrapping half characters correctly

+++ src/main/java/io/karma/pda/api/client/render/graphics/FontAtlas.java
@@ -33,15 +33,15 @@ public interface FontAtlas extends Disposable {
 
     int getSpriteBorder();
 
-    float getMaxGlyphWidth();
+    int getMaxGlyphWidth();
 
-    float getMaxGlyphHeight();
+    int getMaxGlyphHeight();
 
-    float getMaxGlyphBearingX();
+    int getMaxGlyphBearingX();
 
-    float getMaxGlyphBearingY();
+    int getMaxGlyphBearingY();
 
-    float getLineHeight();
+    int getLineHeight();
 
     int getTextureId();
